% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize_first_char.R
\name{capitalize_first_char}
\alias{capitalize_first_char}
\title{Convert the first character to upper case}
\usage{
capitalize_first_char(s)
}
\arguments{
\item{s}{a string}
}
\description{
Convert the first character to upper case
}
\author{
Richel J.C. Bilderbeek
}
