% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_download_url.R
\name{check_download_url}
\alias{check_download_url}
\title{Check if the download URL is valid}
\usage{
check_download_url(
  download_url = get_netmhc2pan_url(),
  verbose = FALSE,
  netmhc2pan_archive_filename = get_netmhc2pan_archive_filename(),
  temp_local_file = tempfile(pattern = "netmhc2pan_check_download_url_")
)
}
\arguments{
\item{download_url}{NetMHCIIpan download URL, for example,
\code{https://www.cbs.dtu.dk/download/12345678-1234-1234-1234-123456789ABC}}

\item{verbose}{set to TRUE for more output}

\item{netmhc2pan_archive_filename}{the NetMHC2pan archive
filename, for example \code{netMHCIIpan-3.2.Linux.tar.gz},
which is used in installation of NetMHC2pan.
Use \link{get_netmhc2pan_archive_filename} to get this filename.}

\item{temp_local_file}{path to the temporary file to store the URL to.
This file will be deleted afterwards.}
}
\description{
Check if the download URL is valid, will stop if not.
The download link expires after 4 hours.
}
