% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_netmhc2pan_installation.R
\name{check_netmhc2pan_installation}
\alias{check_netmhc2pan_installation}
\title{Checks the installation of NetMHC2pan.
Throws a helpful error message if incomplete, 
else does nothing}
\usage{

  check_netmhc2pan_installation(folder_name = get_default_netmhc2pan_folder())
}
\arguments{
\item{folder_name}{the folder to install NetMHC2pan, which is
\code{"/home/[usename]/.local/share"} by default}
}
\value{
Nothing
}
\description{
Checks the installation of NetMHC2pan.
Throws a helpful error message if incomplete, 
else does nothing
}
\author{
Richel J.C. Bilderbeek
}
