% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_os.R
\name{check_os}
\alias{check_os}
\title{Checks if the operating system is supported}
\usage{
check_os(os)
}
\arguments{
\item{os}{the operating system as obtained by \code{rappdirs::app_dir()$os}.
\code{netmhc2pan} supports Linux ('\code{unix}') only}
}
\value{
nothing. Will stop if the OS is unsupported
}
\description{
Checks if the operating system is supported
}
\author{
Richel J.C. Bilderbeek
}
