% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_temp_fasta_filename.R,
%   R/create_temp_xls_filename.R
\name{create_temp_xls_filename}
\alias{create_temp_xls_filename}
\title{Create a \code{.fasta} file}
\usage{
create_temp_xls_filename()

create_temp_xls_filename()
}
\description{
Create a \code{.fasta} file

Create a \code{.xls} file
}
