% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_netmhc2pan_archive_filename.R
\name{get_netmhc2pan_archive_filename}
\alias{get_netmhc2pan_archive_filename}
\title{Get the filename of the NetMHC2pan archive file}
\usage{
get_netmhc2pan_archive_filename()
}
\description{
Get the filename of the NetMHC2pan archive file
}
