% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_download_url.R
\name{get_netmhc2pan_url}
\alias{get_netmhc2pan_url}
\title{Get a NetMHC2pan download URL
Will stop if the NetMHC2pan download URL is invalid}
\usage{
get_netmhc2pan_url()
}
\value{
a download URL
}
\description{
Get a NetMHC2pan download URL
Will stop if the NetMHC2pan download URL is invalid
}
\author{
Richel J.C. Bilderbeek
}
