% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_netmhc2pan.R
\name{install_netmhc2pan}
\alias{install_netmhc2pan}
\title{Install NetMHC2pan to a local folder}
\usage{
install_netmhc2pan()
}
\description{
Install NetMHC2pan to a local folder
}
\author{
Richel J.C. Bilderbeek
}
