% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_netmhc2pan_installed.R
\name{is_netmhc2pan_bin_installed}
\alias{is_netmhc2pan_bin_installed}
\title{Measure if NetMHCIIpan binary is installed locally}
\usage{
is_netmhc2pan_bin_installed(folder_name = get_default_netmhc2pan_folder())
}
\arguments{
\item{folder_name}{the folder to install NetMHCIIpan, which is
\code{"/home/[usename]/.local/share"} by default}
}
\value{
TRUE is NetMHCIIpan binary is installed locally,
  FALSE otherwise
}
\description{
Measure if NetMHCIIpan binary is installed locally
}
\author{
Richèl J.C. Bilderbeek
}
