% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_netmhc2pan_installed.R
\name{is_netmhc2pan_set_up}
\alias{is_netmhc2pan_set_up}
\title{Measure if NetMHCIIpan is set up}
\usage{
is_netmhc2pan_set_up(folder_name = get_default_netmhc2pan_folder())
}
\arguments{
\item{folder_name}{the folder to install NetMHCIIpan, which is
\code{"/home/[usename]/.local/share"} by default}
}
\value{
TRUE is NetMHCIIpan is set up locally,
  FALSE otherwise
}
\description{
Measure if NetMHCIIpan is set up
}
\author{
Richel J.C. Bilderbeek
}
