% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netmhc2pan.R
\docType{package}
\name{netmhc2pan}
\alias{netmhc2pan}
\title{netmhc2pan: predict peptide binfing strengths using NetMHC2pan}
\examples{
if (is_netmhc2pan_installed()) {

  predict_ic50(
    peptides = c("AIAACAMLLV", "ALVCYIVMPV"),
    mhc_haplotype = "DRB1_0416"
  )
}
}
\author{
Richèl J.C. Bilderbeek
}
