% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_netmhc2pan_xls_output.R
\name{read_netmhc2pan_xls_output}
\alias{read_netmhc2pan_xls_output}
\title{Read MHC2pan output}
\usage{
read_netmhc2pan_xls_output(xls_filename)
}
\arguments{
\item{xls_filename}{name of an \code{xls} file}
}
\value{
a table
}
\description{
Read the output produced by MHC2pan,
which is an \code{XLS} file by default.
}
\examples{
xls_filename <- system.file("extdata", "example.xls", package = "netmhc2pan")
read_netmhc2pan_xls_output(xls_filename)
}
