% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(alleles, do_filter, download_url, fasta_filename,
  folder_name, os, temp_xls_filename, verbose)
}
\arguments{
\item{alleles}{one or more alleles, e.g. \code{DRB1_0101}. 
See \link{get_netmhc2pan_alleles} for a full list}

\item{do_filter}{set to TRUE if the results of NetMHCIIpan must be filtered}

\item{download_url}{NetMHCIIpan download URL, for example,
\code{http://www.cbs.dtu.dk/download/12345678-1234-1234-1234-123456789ABC}}

\item{fasta_filename}{the name of a FASTA file with protein sequences}

\item{folder_name}{the folder to install NetMHCIIpan, which is
\code{"/home/[usename]/.local/share"} by default}

\item{os}{the operating system as obtained by \code{rappdirs::app_dir()$os}.
\code{netmhc2pan} supports Linux ('\code{unix}') only}

\item{temp_xls_filename}{name for a temporary \code{xls} filename}

\item{verbose}{set to TRUE for more output}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
  \code{@noRd}. This is not done, as this will disallow all
  functions to find the documentation parameters
}
\author{
Richel J.C. Bilderbeek
}
