% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_netmhc2pan_bin_path.R
\name{get_default_netmhc2pan_bin_path}
\alias{get_default_netmhc2pan_bin_path}
\title{Get the full path to the default NetMHC2pan binary}
\usage{
get_default_netmhc2pan_bin_path()
}
\value{
the full path to the default NetMHC2pan binary
}
\description{
Get the full path to the default NetMHC2pan binary
}
\examples{
get_default_netmhc2pan_bin_path()
}
\author{
Richèl J.C. Bilderbeek
}
