% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_netmhc2pan_folder.R
\name{get_default_netmhc2pan_folder}
\alias{get_default_netmhc2pan_folder}
\title{Get the path to the folder where this package installs
NetMHC2pan by default}
\usage{
get_default_netmhc2pan_folder()
}
\value{
the path to the folder where this package installs
  NetMHC2pan by default
}
\description{
Get the path to the folder where this package installs
NetMHC2pan by default
}
\examples{
  if (rappdirs::app_dir()$os == "unix") {
    testit::assert(
      grepl(
        "/home/[A-Za-z0-9_]*/.local/share",
        get_default_netmhc2pan_folder()
      )
    )
  }
}
\author{
Richel J.C. Bilderbeek
}
