% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_netmhc2pan_subfolder.R
\name{get_default_netmhc2pan_subfolder}
\alias{get_default_netmhc2pan_subfolder}
\title{Get the full path to the NetMHC2pan subfolder}
\usage{
get_default_netmhc2pan_subfolder()
}
\value{
the full path to the NetMHC2pan subfolder
}
\description{
Get the full path to the NetMHC2pan subfolder
}
\author{
Richel J.C. Bilderbeek
}
