% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_netmhc2pan_url.R
\name{get_netmhc2pan_url}
\alias{get_netmhc2pan_url}
\title{Get a NetMHCIIpan download URL
Will stop if the NetMHCIIpan download URL is invalid}
\usage{
get_netmhc2pan_url()
}
\value{
a download URL
}
\description{
Get a NetMHCIIpan download URL
Will stop if the NetMHCIIpan download URL is invalid
}
\author{
Richel J.C. Bilderbeek
}
