% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_netmhc2pan.R
\name{install_netmhc2pan}
\alias{install_netmhc2pan}
\title{Install NetMHCIIpan}
\usage{
install_netmhc2pan(
  download_url = get_netmhc2pan_url(),
  verbose = FALSE,
  netmhc2pan_archive_filename = get_netmhc2pan_archive_filename(),
  netmhc2pan_folder_name = get_default_netmhc2pan_folder(),
  temp_local_file = tempfile(pattern = "netmhc2pan_install_netmhc2pan_")
)
}
\arguments{
\item{download_url}{NetMHCIIpan download URL, for example,
\code{https://www.cbs.dtu.dk/download/12345678-1234-1234-1234-123456789ABC}}

\item{verbose}{set to TRUE for more output}

\item{netmhc2pan_archive_filename}{the NetMHC2pan archive
filename, for example \code{netMHCIIpan-3.2.Linux.tar.gz},
which is used in installation of NetMHC2pan.
Use \link{get_netmhc2pan_archive_filename} to get this filename.}

\item{netmhc2pan_folder_name}{the folder (to be) used by NetMHCIIpan.
Use \link{get_default_netmhc2pan_folder} to see the location
of the default NetMHCIIpan folder}

\item{temp_local_file}{path to the temporary file to store the URL to.
This file will be deleted afterwards.}
}
\value{
Nothing
}
\description{
Install NetMHCIIpan to a local folder.
}
\examples{
\dontrun{
  install_netmhc2pan()
}
}
\author{
Richèl J.C. Bilderbeek
}
