% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_netmhc2pan.R
\name{install_netmhc2pan}
\alias{install_netmhc2pan}
\title{Install NetMHCIIpan to a local folder}
\usage{
install_netmhc2pan(download_url = get_netmhc2pan_url(),
  folder_name = rappdirs::user_data_dir())
}
\arguments{
\item{download_url}{NetMHCIIpan download URL, for example,
\code{http://www.cbs.dtu.dk/download/12345678-1234-1234-1234-123456789ABC}}

\item{folder_name}{the folder to install NetMHCIIpan, which is
\code{"/home/[usename]/.local/share"} by default}
}
\description{
Install NetMHCIIpan to a local folder
}
\examples{
  if (!is_netmhc2pan_installed()) install_netmhc2pan()
  testit::assert(is_netmhc2pan_installed())
}
\author{
Richel J.C. Bilderbeek
}
