% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_netmhc2pan.R
\name{install_netmhc2pan_bin}
\alias{install_netmhc2pan_bin}
\title{Install the NetMHCIIpan binary to a local folder}
\usage{
install_netmhc2pan_bin(
  download_url = get_netmhc2pan_url(),
  netmhc2pan_folder_name = get_default_netmhc2pan_folder()
)
}
\arguments{
\item{download_url}{NetMHCIIpan download URL, for example,
\code{http://www.cbs.dtu.dk/download/12345678-1234-1234-1234-123456789ABC}}

\item{netmhc2pan_folder_name}{the folder (to be) used by NetMHCIIpan.
Use \link{get_default_netmhc2pan_folder} to see the location
of the default NetMHCIIpan folder}
}
\description{
Install the NetMHCIIpan binary to a local folder
}
\author{
Richèl J.C. Bilderbeek
}
