% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_netmhc2pan.R
\name{install_netmhc2pan_bin}
\alias{install_netmhc2pan_bin}
\title{Install the NetMHC2pan binary to a local folder}
\usage{
install_netmhc2pan_bin(download_url = get_netmhc2pan_url(),
  folder_name = rappdirs::user_data_dir(), verbose = FALSE)
}
\arguments{
\item{download_url}{NetMHC2pan download URL, for example,
\code{http://www.cbs.dtu.dk/download/12345678-1234-1234-1234-123456789ABC}}

\item{folder_name}{the folder to install NetMHC2pan, which is
\code{"/home/[usename]/.local/share"} by default}

\item{verbose}{set to TRUE for more output}
}
\description{
Install the NetMHC2pan binary to a local folder
}
\author{
Richel J.C. Bilderbeek
}
