% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_netmhc2pan.R
\name{install_netmhc2pan_data}
\alias{install_netmhc2pan_data}
\title{Install the NetMHCIIpan binary to a local folder}
\usage{
install_netmhc2pan_data(netmhc2pan_folder_name = rappdirs::user_data_dir())
}
\arguments{
\item{netmhc2pan_folder_name}{the folder (to be) used by NetMHCIIpan.
Use \link{get_default_netmhc2pan_folder} to see the location
of the default NetMHCIIpan folder}
}
\description{
Install the NetMHCIIpan binary to a local folder
}
\author{
Richèl J.C. Bilderbeek
}
