% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_netmhc2pan_data.R
\name{install_netmhc2pan_data}
\alias{install_netmhc2pan_data}
\title{Install the NetMHCIIpan data to a local folder}
\usage{
install_netmhc2pan_data(
  netmhc2pan_folder_name = rappdirs::user_data_dir(),
  verbose = FALSE
)
}
\arguments{
\item{netmhc2pan_folder_name}{the folder (to be) used by NetMHCIIpan.
From this location, a subfolder for NetMHCIIpan is created.
Use \link{get_default_netmhc2pan_folder} to see the location
of the default NetMHCIIpan folder.
Use \link{get_default_netmhc2pan_subfolder} to see the location
of the default NetMHCIIpan subfolder.}

\item{verbose}{set to TRUE for more output}
}
\value{
Nothing
}
\description{
Install the NetMHCIIpan data to a local folder
}
\examples{
\dontrun{
  install_netmhc2pan_data()
}
}
\author{
Richèl J.C. Bilderbeek
}
