% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_netmhc2pan_installed.R
\name{is_netmhc2pan_installed}
\alias{is_netmhc2pan_installed}
\title{Measure if NetMHC2pan is installed locally}
\usage{
is_netmhc2pan_installed(folder_name = get_default_netmhc2pan_folder(),
  os = rappdirs::app_dir()$os)
}
\value{
TRUE is NetMHC2pan is installed locally,
  FALSE otherwise
}
\description{
Measure if NetMHC2pan is installed locally
}
\author{
Richel J.C. Bilderbeek
}
