% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_netmhc2pan_xls_output.R
\name{read_netmhc2pan_xls_output}
\alias{read_netmhc2pan_xls_output}
\title{Read MHC2pan output}
\usage{
read_netmhc2pan_xls_output(xls_filename)
}
\description{
Read the output produced by MHC2pan,
which is an XLS file by default.
}
