% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_netmhc2pan.R
\name{run_netmhc2pan}
\alias{run_netmhc2pan}
\title{Run NetMHC2pan}
\usage{
run_netmhc2pan(fasta_filename, do_filter = FALSE)
}
\arguments{
\item{fasta_filename}{the name of a FASTA file with protein sequences}

\item{do_filter}{set to TRUE if the results of NetMHC2pan must be filtered}
}
\value{
a data frame with the NetMHC2pan results
}
\description{
Run NetMHC2pan
}
\author{
Richel J.C. Bilderbeek
}
