% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_netmhc2pan.R
\name{run_netmhc2pan}
\alias{run_netmhc2pan}
\title{Run NetMHC2pan}
\usage{
run_netmhc2pan(fasta_filename, alleles = "DRB1_0101",
  folder_name = get_default_netmhc2pan_folder(),
  temp_xls_filename = tempfile(fileext = ".xls"))
}
\arguments{
\item{fasta_filename}{the name of a FASTA file with protein sequences}

\item{alleles}{one or more alleles, e.g. \code{DRB1_0101}. 
See \link{get_netmhc2pan_alleles} for a full list}

\item{folder_name}{the folder to install NetMHC2pan, which is
\code{"/home/[usename]/.local/share"} by default}

\item{temp_xls_filename}{name for a temporary \code{xls} filename}
}
\value{
a data frame with the NetMHC2pan results
}
\description{
Run NetMHC2pan
}
\examples{
  testit::assert(is_netmhc2pan_installed())
  
  fasta_filename <- system.file(
    "extdata", "example.fasta", package = "netmhc2pan"
  )
  
  # One default allele
  df <- run_netmhc2pan(fasta_filename)
  testit::assert(
    all(
      colnames(df) == 
      c("Pos", "Peptide", "ID", "Allele", "one_minus_log50k", "nM", "Rank")
    )
  )
  testit::assert(nrow(df) == 9)
  testit::assert(all(df$Allele == "DRB1_0101"))
  
  # Two alleles
  alleles <- c("DRB1_0101", "DRB1_0102")
  # Alleles must be in NetMHC2pan
  testit::assert(all(alleles \%in\% get_netmhc2pan_alleles()))
  # Run NetMHCpan with these two alleles
  df <- run_netmhc2pan(fasta_filename, alleles = alleles)
  testit::assert(nrow(df) == 18)
}
\author{
Richel J.C. Bilderbeek
}
