% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_netmhc2pan.R
\name{set_up_netmhc2pan}
\alias{set_up_netmhc2pan}
\title{Install the NetMHC2pan binary to a local folder}
\usage{
set_up_netmhc2pan(folder_name = rappdirs::user_data_dir(),
  verbose = FALSE, os = rappdirs::app_dir()$os)
}
\arguments{
\item{folder_name}{the folder to install NetMHC2pan, which is
\code{"/home/[usename]/.local/share"} by default}

\item{verbose}{set to TRUE for more output}

\item{os}{the operating system as obtained by \code{rappdirs::app_dir()$os}.
\code{netmhc2pan} supports Linux ('\code{unix}') only}
}
\description{
Install the NetMHC2pan binary to a local folder
}
\author{
Richel J.C. Bilderbeek
}
