% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_netmhc2pan.R
\name{set_up_netmhc2pan}
\alias{set_up_netmhc2pan}
\title{Install the NetMHCIIpan binary to a local folder}
\usage{
set_up_netmhc2pan(netmhc2pan_folder_name = get_default_netmhc2pan_folder())
}
\arguments{
\item{netmhc2pan_folder_name}{the folder (to be) used by NetMHCIIpan.
Use \link{get_default_netmhc2pan_folder} to see the location
of the default NetMHCIIpan folder}
}
\value{
Nothing
}
\description{
Install the NetMHCIIpan binary to a local folder
}
\examples{
\dontrun{
  set_up_netmhc2pan()
}
}
\author{
Richèl J.C. Bilderbeek
}
