% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_netmhc2pan.R
\name{set_up_netmhc2pan}
\alias{set_up_netmhc2pan}
\title{Install the NetMHCIIpan binary to a local folder}
\usage{
set_up_netmhc2pan(folder_name = rappdirs::user_data_dir())
}
\arguments{
\item{folder_name}{the folder to install NetMHCIIpan, which is
\code{"/home/[usename]/.local/share"} by default}
}
\description{
Install the NetMHCIIpan binary to a local folder
}
\author{
Richèl J.C. Bilderbeek
}
