% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uninstall_netmhc2pan.R
\name{uninstall_netmhc2pan}
\alias{uninstall_netmhc2pan}
\title{Uninstall NetMHCIIpan}
\usage{
uninstall_netmhc2pan(folder_name = rappdirs::user_data_dir())
}
\arguments{
\item{folder_name}{name of the folder
where the NetMHCIIpan files are installed.
The name of the NetMHCIIpan binary file will be at
\code{[folder_name]/[subfolder_basename]/netMHCIIpan}}
}
\description{
Uninstall NetMHCIIpan
}
\author{
Richel J.C. Bilderbeek
}
